requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.31.0
requests==2.31.0
langchain==0.1.11
langchain-openai==0.0.8
langgraph==0.0.26
pydantic==2.5.3
pandas==2.2.0
numpy==1.26.3
chromadb==0.4.24
tiktoken==0.5.2
python-dotenv==1.0.0
sentence-transformers>=3.0.1
transformers>=4.38.0
torch>=2.3,<2.6
PyPDF2==3.0.1
python-docx==1.1.0